requirements.txt
pyarrow~=11.0.0
psutil==5.9.4
dash-bootstrap-components==1.4.0
dash~=2.11.1
dash-mantine-components==0.12.1
diskcache==5.4.0
dash-daq==0.5.0
pymongo~=4.3.3
dnspython
pandas~=1.5.3
networkx~=3.0
pytz~=2022.7.1
python-dotenv~=0.19.2
requests~=2.31.0
osmnx~=1.3.0
ortools>=9.2.9972 #import for compatibility with anytime algo
geopandas
numpy~=1.24.2
Shapely~=2.0.1
pyproj~=3.4.1
plotly~=5.15.0
mapbox~=0.18.1
phonenumbers~=8.13.7
python-dateutil~=2.8.2
openpyxl
memoization
dateparser
dash_extensions
dash_leaflet
pdoc3~=0.10.0
cython
names
dash-auth
google-cloud-secret-manager
gtfs_functions~=2.0.3
celery
redis
geopy

